move cloud shelf code as a pre step
